requirements.txt
google-genai
requests
python-dotenv
duckduckgo_search
